Game of War Card Game
A simple C# project that uses enums, classes, random variables.   The program demonstrates basic programming methodology concepts such as looping using foreach over a enumerated data types and a simple while loop.  The "cards" are stored in a list and shuffled using a randomly generated number in the range of 0 to 51.